+++ tests/posit/api/quire_accumulation.cpp
@@ -20,18 +20,18 @@
 //#define ISSUE_45_DEBUG
 #ifdef ISSUE_45_DEBUG
 // forward reference
-template<size_t nbits, size_t es, size_t capacity> void Issue45_2();
+template<unsigned nbits, unsigned es, unsigned capacity> void Issue45_2();
 #endif
 
-template<size_t nbits, size_t es>
+template<unsigned nbits, unsigned es>
 void PrintTestVector(std::ostream& ostr, const std::vector< sw::universal::posit<nbits,es> >& pv) {
 	std::for_each (begin(pv), end(pv), [&ostr](const sw::universal::posit<nbits,es>& p){
 		ostr << p << std::endl;
 	});
 }
 
-template<size_t nbits, size_t es, size_t capacity>
-int GenerateQuireAccumulationTestCase(bool bReportIndividualTestCases, size_t nrOfElements, const sw::universal::posit<nbits,es>& seed) {
+template<unsigned nbits, unsigned es, unsigned capacity>
+int GenerateQuireAccumulationTestCase(bool bReportIndividualTestCases, unsigned nrOfElements, const sw::universal::posit<nbits,es>& seed) {
 	using namespace sw::universal;
 	int nrOfFailedTestCases = 0;
 	std::stringstream ss;
@@ -44,7 +44,7 @@ int GenerateQuireAccumulationTestCase(bool bReportIndividualTestCases, size_t nr
 // initialize a vector
 template<typename Vector, typename Scalar>
 void init(Vector& x, const Scalar& value) {
-	for (size_t i = 0; i < x.size(); ++i) x[i] = value;
+	for (unsigned i = 0; i < x.size(); ++i) x[i] = value;
 }
 
 // regular dot product
@@ -55,13 +55,13 @@ typename Vector::value_type dot(const Vector& a, const Vector& b) {
 		std::cerr << "vectors are not the same size\n";
 		return sum;
 	}
-	for (size_t i = 0; i < size(a); ++i) {
+	for (unsigned i = 0; i < size(a); ++i) {
 		sum += a[i] * b[i];
 	}
 	return sum;
 }
 
-template<size_t nbits, size_t es, size_t nrElements = 16>
+template<unsigned nbits, unsigned es, unsigned nrElements = 16>
 int ValidateExactDotProduct() {
 	using namespace sw::universal;
 	int nrOfFailures = 0;
@@ -116,7 +116,7 @@ int ValidateQuireMagnitudeComparison() {
 	return 0;
 }
 
-template<size_t nbits, size_t es, size_t capacity = 2>
+template<unsigned nbits, unsigned es, unsigned capacity = 2>
 int ValidateSignMagnitudeTransitions() {
 	using namespace sw::universal;
 
@@ -237,17 +237,17 @@ int ValidateSignMagnitudeTransitions() {
 	return nrOfFailedTestCases;
 }
 
-template<size_t nbits, size_t es, size_t capacity = 2>
+template<unsigned nbits, unsigned es, unsigned capacity = 2>
 int ValidateCarryPropagation() {
 	using namespace sw::universal;
 	int nrOfFailedTests = 0;
 
-	constexpr size_t mbits = 2 * (nbits - 2 - es);
+	constexpr unsigned mbits = 2 * (nbits - 2 - es);
 	quire<nbits, es, capacity> q;
 	posit<nbits, es> mp(SpecificValue::minpos);
 	internal::value<mbits> minpos_square = quire_mul(mp, mp);
-	constexpr size_t NR_INCREMENTS_TO_OVERFLOW = (size_t(1) << (q.qbits+1));
-	for (size_t i = 0; i < NR_INCREMENTS_TO_OVERFLOW; ++i) {
+	constexpr unsigned NR_INCREMENTS_TO_OVERFLOW = (unsigned(1) << (q.qbits+1));
+	for (unsigned i = 0; i < NR_INCREMENTS_TO_OVERFLOW; ++i) {
 		q += minpos_square;
 	}
 	std::cout << q << '\n';
@@ -256,19 +256,19 @@ int ValidateCarryPropagation() {
 	return nrOfFailedTests;
 }
 
-template<size_t nbits, size_t es, size_t capacity = 2>
+template<unsigned nbits, unsigned es, unsigned capacity = 2>
 int ValidateBorrowPropagation() {
 	using namespace sw::universal;
 	int nrOfFailedTests = 0;
 
-	constexpr size_t mbits = 2 * (nbits - 2 - es);
+	constexpr unsigned mbits = 2 * (nbits - 2 - es);
 	quire<nbits, es, capacity> q;
 	posit<nbits, es> mp(SpecificValue::minpos);
 	internal::value<mbits> minpos_square = quire_mul(mp, mp);
 	q -= minpos_square;
 	std::cout << q << '\n';
-	constexpr size_t NR_DECREMENTS_TO_OVERFLOW = (size_t(1) << (q.qbits + 1));
-	for (size_t i = 0; i < NR_DECREMENTS_TO_OVERFLOW-1; ++i) {
+	constexpr unsigned NR_DECREMENTS_TO_OVERFLOW = (unsigned(1) << (q.qbits + 1));
+	for (unsigned i = 0; i < NR_DECREMENTS_TO_OVERFLOW-1; ++i) {
 		q -= minpos_square;
 	}
 	std::cout << q << '\n';
@@ -277,7 +277,7 @@ int ValidateBorrowPropagation() {
 	return nrOfFailedTests;
 }
 
-template<size_t nbits, size_t es, size_t capacity = 2>
+template<unsigned nbits, unsigned es, unsigned capacity = 2>
 int ValidateQuireAccumulation(bool bReportIndividualTestCases) {
 	int nrOfFailedTests = 0;
 
@@ -426,11 +426,11 @@ catch (...) {
 // specific debug scenarios of note
 //
 // use forward reference to bring them up to the main body
-// template<size_t nbits, size_t es, size_t capacity> void Issue45();
-// template<size_t nbits, size_t es, size_t capacity> void Issue45_2();
+// template<unsigned nbits, unsigned es, unsigned capacity> void Issue45();
+// template<unsigned nbits, unsigned es, unsigned capacity> void Issue45_2();
 
 // test case for github issue #45
-template<size_t nbits, size_t es>
+template<unsigned nbits, unsigned es>
 void Issue45() {
 	using ScalarType = sw::universal::posit<nbits, es>;
 	using magnitude = sw::universal::posit<nbits, es>;
@@ -506,13 +506,13 @@ taking -2.68435e+08 += quire_mul(-0.00828552, 0.000999451) (which equals -8.2809
 */
 
 // step by step testing to find where the failure occurred
-template<size_t nbits, size_t es, size_t capacity = 30>
+template<unsigned nbits, unsigned es, unsigned capacity = 30>
 void Issue45_2() {
 	using namespace sw::universal;
 
 	std::cout << "Debug of issue #45\n";
 
-	constexpr size_t mbits = 2 * (nbits - 2 - es);
+	constexpr unsigned mbits = 2 * (nbits - 2 - es);
 	sw::universal::quire<nbits, es, capacity> q, q_base;
 	sw::universal::value<mbits> unrounded, q_value;
 	sw::universal::bitblock<mbits> fraction;
@@ -661,13 +661,13 @@ void Issue45_2() {
 	std::cout << q << std::endl;
 
 	{
-		constexpr size_t mbits = 2 * (nbits - 2 - es);
+		constexpr unsigned mbits = 2 * (nbits - 2 - es);
 		sw::universal::quire<nbits, es, capacity> q, q_base;
 
 		// inefficient as we are copying a whole quire just to reset the sign bit, but we are leveraging the comparison logic
 		//quire<nbits, es, capacity> absq = abs(*this);
-		//constexpr size_t qbits = (size_t(1) << es) * (4 * nbits - 8) + capacity;
-		//constexpr size_t fbits = nbits - 3 - es;
+		//constexpr unsigned qbits = (unsigned(1) << es) * (4 * nbits - 8) + capacity;
+		//constexpr unsigned fbits = nbits - 3 - es;
 		//value<qbits> absq = abs(q);
 		quire <nbits, es, capacity> absq = abs(q);
 		internal::value<mbits> absv = abs(unrounded);
